Add support for dynamic macro loading
Modify the loadMacro interface so that the class name is passed
as well as the dex bytecode: this is necessary for loading the
class at runtime.

The implementation still needs to first write the bytecode to a
local file, then load the actual macro class. It would improve
both lisibility and reliability to perform only from memory
representation.

See http://android-developers.blogspot.fr/2011/07/custom-class-loading-in-dalvik.html
for a great example of dynamic class loading.